bot/bot_api.py
from linebot import LineBotApi, WebhookHandler

from LineBotBackend import settings

line_bot_api = LineBotApi(settings.LINE_CHANNEL_ACCESS_TOKEN)
handler = WebhookHandler(settings.LINE_CHANNEL_SECRET)